fix:设置主页面不可旋转，将小组件刷新逻辑调整到课表页面
