+++ app/src/main/AndroidManifest.xml
@@ -193,7 +193,8 @@
             android:screenOrientation="portrait" />
         <activity
             android:name=".core.MainActivity"
-            android:exported="true">
+            android:exported="true"
+            android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
