+++ app/src/main/java/com/example/changli_planet_app/common/redux/store/UserStore.kt
@@ -312,7 +312,7 @@ class UserStore : Store<UserState, UserAction>() {
 
 
                                     }
-                                    PlanetApplication.clearCacheAll()
+                                    PlanetApplication.clearSchoolDataCacheAll()
                                 }
 
                                 else -> {
