+++ app/src/main/java/com/example/changli_planet_app/core/PlanetApplication.kt
@@ -45,11 +45,19 @@ class PlanetApplication : Application() {
 
         fun clearCacheAll() {
             CoroutineScope(Dispatchers.IO).launch {
-//                accessToken = ""
-//                MMKV.mmkvWithID("import_cache").clearAll()
-//                MMKV.mmkvWithID("content_cache").clearAll()
-//                MMKV.mmkvWithID(TIME_TABLE_APP_WIDGET).clearAll()
-//                CoursesDataBase.getDatabase(appContext).courseDao().clearAllCourses()
+                accessToken = ""
+                MMKV.mmkvWithID("import_cache").clearAll()
+                MMKV.mmkvWithID("content_cache").clearAll()
+                MMKV.mmkvWithID(TIME_TABLE_APP_WIDGET).clearAll()
+                CoursesDataBase.getDatabase(appContext).courseDao().clearAllCourses()
+            }
+        }
+        fun clearSchoolDataCacheAll(){
+            CoroutineScope(Dispatchers.IO).launch {
+                MMKV.mmkvWithID("import_cache").clearAll()
+                MMKV.mmkvWithID("content_cache").clearAll()
+                MMKV.mmkvWithID(TIME_TABLE_APP_WIDGET).clearAll()
+                CoursesDataBase.getDatabase(appContext).courseDao().clearAllCourses()
             }
         }
